+++ p.sh
@@ -1,0 +1,1 @@
+flutter packages pub publish -v